changed header, fixed styles
